requirements.txt
streamlit==1.32.0
gspread==6.0.0
google-auth==2.28.2
requests==2.31.0